+++ main.go
@@ -196,11 +196,11 @@ func main() {
 		os.Exit(1)
 	}
 
-	if err = (&controllers.ConfigMapReconciler{
+	if err = (&controllers.CrossplaneClusterConfigReconciler{
 		Client:                mgr.GetClient(),
 		BaseDomain:            workloadClusterBaseDomain,
 		ManagementClusterName: managementClusterName,
-	}).SetupWithManager(mgr); err != nil {
+	}).SetupWithManager(ctx, mgr); err != nil {
 		setupLog.Error(err, "unable to create controller", "controller", "Frigate")
 		os.Exit(1)
 	}